changes elapsed to a float adds time member to Testcase class and marks it as an XmlAttribute

+++ src/main/java/testrail/testrail/TestRailObjects/Result.java
@@ -24,10 +24,10 @@ package testrail.testrail.TestRailObjects;
 public class Result {
     private int caseId;
     private int statusId;
-    private int elapsed;
+    private float elapsed;
     private String comment;
 
-    public Result(int caseId, int statusId, int elapsed, String comment,) {
+    public Result(int caseId, int statusId, float elapsed, String comment,) {
         this.caseId = caseId;
         this.statusId = statusId;
         this.comment = comment;
@@ -36,11 +36,11 @@ public class Result {
 
     public void setCaseId(int caseId) { this.caseId = caseId; }
     public void setStatusId(int statusId) { this.statusId = statusId; }
-    public void setElapsed(int timeInSeconds) { this.elapsed = timeInSeconds; }
+    public void setElapsed(float timeInSeconds) { this.elapsed = timeInSeconds; }
     public void setComment(String comment) { this.comment = comment; }
 
     public int getCaseId() { return this.caseId; }
     public int getStatusId() { return this.statusId; }
-    public int getElapsed() { return this.elapsed; }
+    public float getElapsed() { return this.elapsed; }
     public String getComment() { return this.comment; }
 }
